requirements.txt
python = "^3.13"
django = "^5.1.4"
django-filter = "^24.3"
djangorestframework = "^3.15.2"
djoser = "^2.3.1"
djangorestframework-simplejwt = "^5.3.1"
psycopg = "^3.2.3"
django-debug-toolbar = "^4.4.6"